requirements.txt
# Trading Bot Requirements
# Core dependencies
openai>=1.0.0
requests>=2.28.0
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0
pytz>=2023.3
pandas>=1.5.0
numpy>=1.22.0

# For Windows Service (includes servicemanager, win32event, win32service, win32serviceutil)
pywin32>=305; platform_system=="Windows"

# For data analysis
matplotlib>=3.6.0
seaborn>=0.12.0

# For logging and error handling
logging>=0.5.1.2